+++ backend/rest_app/fast_api_dependencies/voyago_client.py
@@ -1,0 +1,16 @@
+"""voyago client dependency"""
+
+from backend.app.completions.client import AIClient
+from backend.app.services.unsplash.unsplash_service import UnsplashService
+from backend.app.voyago import Voyago
+
+
+def get_voyago() -> Voyago:
+    """gets an initialized voyago object
+
+    injected with client and service
+    """
+    client = AIClient()
+    unsplash = UnsplashService()
+    voyago = Voyago(client=client, unsplash=unsplash)
+    return voyago
