+++ backend/rest_app/routers/itinerary.py
@@ -5,7 +5,7 @@ from starlette.exceptions import HTTPException
 
 from backend.app.exceptions import TripPlanGenerationError
 from backend.app.models.recommendations import RecommendationQuery, VisualItinerary
-from backend.rest_app.dependencies.voyago_client import get_voyago
+from backend.rest_app.fast_api_dependencies.voyago_client import get_voyago
 
 router = APIRouter(prefix="/itinerary", tags=["itinerary"])
